Get feather-server split to compile
