+++ server/chunk/src/lib.rs
@@ -1,4 +1,8 @@
 #![forbid(unsafe_code)]
 
-mod chunk_logic;
-mod chunk_worker;
+mod chunk_manager;
+pub mod chunk_worker;
+mod save;
+
+pub use chunk_manager::*;
+pub use save::*;
