+++ server/chunk/src/save.rs
@@ -1,17 +1,18 @@
 //! Handles saving of chunks and entities
 
-use crate::entity::ComponentSerializer;
-use crate::game::Game;
-use crate::p_inventory::EntityInventory;
-use crate::{chunk_logic, TICK_TIME, TPS};
-use feather_core::entity::BaseEntityData;
-use feather_core::player_data::{InventorySlot, PlayerData};
-use feather_core::{ChunkPosition, Gamemode, Position, Vec3d};
+use crate::{chunk_manager, ChunkWorkerHandle};
+use feather_core::anvil::entity::BaseEntityData;
+use feather_core::anvil::player::{InventorySlot, PlayerData};
+use feather_core::inventory::Inventory;
+use feather_core::util::{ChunkPosition, Gamemode, Position, Vec3d};
+use feather_server_types::{
+    ChunkLoadEvent, ChunkUnloadEvent, ComponentSerializer, Game, PlayerJoinEvent, PlayerLeaveEvent,
+    Uuid, TICK_LENGTH, TPS,
+};
 use fecs::{Entity, World};
 use std::collections::VecDeque;
 use std::path::Path;
 use std::sync::Arc;
-use uuid::Uuid;
 
 /// A chunk to save + the tick count at which to do so.
 #[derive(Clone, Copy, Debug)]
@@ -24,47 +25,63 @@ struct SaveTask {
 
 /// Queue of chunks to save.
 #[derive(Debug, Default)]
-pub struct SaveQueue(VecDeque<SaveTask>);
+struct SaveQueue(VecDeque<SaveTask>);
 
 /// On a chunk load, adds the chunk to the save queue.
-pub fn on_chunk_load_queue_for_saving(game: &mut Game, chunk: ChunkPosition) {
-    queue_for_saving(game, chunk);
+#[fecs::event_handler]
+pub fn on_chunk_load_queue_for_saving(
+    event: &ChunkLoadEvent,
+    game: &mut Game,
+    #[default] save_queue: &mut SaveQueue,
+) {
+    queue_for_saving(game, save_queue, event.chunk);
 }
 
 /// On a chunk unload, saves the chunk first.
-pub fn on_chunk_unload_save_chunk(game: &mut Game, world: &World, chunk: ChunkPosition) {
-    save_chunk_at(game, world, chunk);
+#[fecs::event_handler]
+pub fn on_chunk_unload_save_chunk(
+    event: &ChunkUnloadEvent,
+    game: &mut Game,
+    world: &mut World,
+    chunk_worker_handle: &ChunkWorkerHandle,
+) {
+    save_chunk_at(game, world, event.chunk, chunk_worker_handle);
 }
 
-fn queue_for_saving(game: &mut Game, chunk: ChunkPosition) {
+fn queue_for_saving(game: &mut Game, save_queue: &mut SaveQueue, chunk: ChunkPosition) {
     let tick_to_save_at =
-        game.tick_count + (game.config.world.save_interval.as_millis() as u64) / TICK_TIME;
+        game.tick_count + (game.config.world.save_interval.as_millis() as u64) / TICK_LENGTH;
 
     let task = SaveTask {
         chunk,
         at: tick_to_save_at,
     };
 
-    game.save_queue.0.push_back(task);
+    save_queue.0.push_back(task);
 }
 
 /// System which checks for chunks which have been queued for saving
 /// and, if it is time, saves them.
 #[fecs::system]
-pub fn chunk_save(game: &mut Game, world: &mut World) {
+pub fn chunk_save(
+    game: &mut Game,
+    world: &mut World,
+    save_queue: &mut SaveQueue,
+    chunk_worker_handle: &ChunkWorkerHandle,
+) {
     // no need to run this system every tick
     if game.tick_count % TPS != 0 {
         return;
     }
 
     loop {
-        let task = match game.save_queue.0.front().copied() {
+        let task = match save_queue.0.front().copied() {
             Some(task) => task,
             None => return, // no save tasks to run
         };
 
         if game.chunk_map.chunk_at(task.chunk).is_none() {
-            game.save_queue
+            save_queue
                 .0
                 .pop_front()
                 .expect("we just verified the front task exists");
@@ -73,22 +90,27 @@ pub fn chunk_save(game: &mut Game, world: &mut World) {
 
         if task.at <= game.tick_count {
             // Save the chunk, then pop the task from the queue.
-            save_chunk_at(game, world, task.chunk);
+            save_chunk_at(game, world, task.chunk, chunk_worker_handle);
 
-            game.save_queue
+            save_queue
                 .0
                 .pop_front()
                 .expect("we just verified the front task exists");
 
             // Requeue the chunk for saving again.
-            queue_for_saving(game, task.chunk);
+            queue_for_saving(game, save_queue, task.chunk);
         } else {
             return;
         }
     }
 }
 
-pub fn save_chunk_at(game: &Game, world: &World, pos: ChunkPosition) {
+pub fn save_chunk_at(
+    game: &Game,
+    world: &World,
+    pos: ChunkPosition,
+    chunk_worker_handle: &ChunkWorkerHandle,
+) {
     let chunk = game
         .chunk_map
         .chunk_handle_at(pos)
@@ -114,22 +136,22 @@ pub fn save_chunk_at(game: &Game, world: &World, pos: ChunkPosition) {
         })
         .collect();
 
-    trace!("Queuing chunk at {} for saving", pos);
-    chunk_logic::save_chunk(
-        &game.chunk_worker_handle,
+    log::trace!("Queuing chunk at {} for saving", pos);
+    chunk_manager::save_chunk(
+        chunk_worker_handle,
         game.chunk_map.chunk_handle_at(pos).unwrap(),
         entities,
     );
 }
 
-pub fn on_player_leave_save_data(game: &Game, world: &World, player: Entity) {
-    save_player_data(game, world, player);
+#[fecs::event_handler]
+pub fn on_player_leave_save_data(event: &PlayerLeaveEvent, game: &Game, world: &mut World) {
+    save_player_data(game, world, event.player);
 }
 
 pub fn save_player_data(game: &Game, world: &World, player: Entity) {
     let inventory = world
-        .get::<EntityInventory>(player)
-        .inventory
+        .get::<Inventory>(player)
         .items()
         .iter()
         .enumerate()
@@ -151,7 +173,7 @@ pub fn save_player_data(game: &Game, world: &World, player: Entity) {
     let config = Arc::clone(&game.config);
 
     game.running_tasks.schedule(async move {
-        match feather_core::player_data::save_player_data(
+        match feather_core::anvil::player::save_player_data(
             &Path::new(&config.world.name),
             uuid,
             &data,
@@ -159,7 +181,7 @@ pub fn save_player_data(game: &Game, world: &World, player: Entity) {
         .await
         {
             Ok(_) => (),
-            Err(e) => error!("Failed to save player data for UUID {}: {}", uuid, e),
+            Err(e) => log::error!("Failed to save player data for UUID {}: {}", uuid, e),
         }
     });
 }
