+++ server/config/Cargo.toml
@@ -7,7 +7,8 @@ edition = "2018"
 [dependencies]
 feather-util = { path = "../../core/util" }
 
-thiserror = "1.0"
+tokio = { version = "0.2", features = ["full"] }
+anyhow = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 humantime-serde = "1.0"
 toml = "0.5"
