+++ server/entity/src/broadcasters.rs
@@ -1,11 +1,13 @@
 mod entity_creation;
 mod entity_deletion;
 mod inventory;
+mod item_collect;
 mod metadata;
 mod movement;
 
 pub use self::inventory::*;
 pub use entity_creation::*;
 pub use entity_deletion::*;
+pub use item_collect::*;
 pub use metadata::*;
 pub use movement::*;
