+++ server/entity/src/broadcasters/item_collect.rs
@@ -1,11 +1,10 @@
-use crate::entity::item::ItemCollectEvent;
-use crate::entity::EntityId;
-use crate::game::Game;
 use feather_core::network::packets::CollectItem;
+use feather_server_types::{EntityId, Game, ItemCollectEvent};
 use fecs::World;
 
 /// Sends `CollectItem` packet when an item is collected.
-pub fn on_item_collect_broadcast(game: &Game, world: &World, event: &ItemCollectEvent) {
+#[fecs::event_handler]
+pub fn on_item_collect_broadcast(event: &ItemCollectEvent, game: &Game, world: &mut World) {
     let packet = CollectItem {
         collected: world.get::<EntityId>(event.item).0,
         collector: world.get::<EntityId>(event.collector).0,
