+++ server/src/systems.rs
@@ -1,34 +1,41 @@
 //! Defines all systems and the order in which they are executed.
 
-use super::*;
 use fecs::Executor;
+
+use feather_server_chunk as chunk_logic;
+use feather_server_entity as entity;
+use feather_server_physics as physics;
+use feather_server_player as player;
+use feather_server_types as game;
+use feather_server_util as util;
+use feather_server_weather as weather;
 
 pub fn build_executor() -> Executor {
     Executor::new()
-        .with(network::poll_player_disconnect)
-        .with(network::poll_new_clients)
+        .with(player::poll_player_disconnect)
+        .with(player::poll_new_clients)
         .with(physics::entity_physics)
-        .with(packet_handlers::handle_movement_packets)
-        .with(packet_handlers::handle_creative_inventory_action)
-        .with(packet_handlers::handle_held_item_change)
-        .with(packet_handlers::handle_animation)
-        .with(packet_handlers::handle_player_block_placement)
-        .with(packet_handlers::handle_player_use_item)
-        .with(packet_handlers::handle_player_digging)
-        .with(packet_handlers::handle_chat)
-        .with(weather::handle_weather)
+        .with(player::handle_movement_packets)
+        .with(player::handle_creative_inventory_action)
+        .with(player::handle_held_item_change)
+        .with(player::handle_animation)
+        .with(player::handle_player_block_placement)
+        .with(player::handle_player_use_item)
+        .with(player::handle_player_digging)
+        .with(player::handle_chat)
+        .with(weather::update_weather)
         .with(entity::item::item_collect)
         .with(chunk_logic::chunk_load)
         .with(chunk_logic::chunk_unload)
         .with(chunk_logic::chunk_optimize)
-        .with(view::check_crossed_chunks)
-        .with(broadcasters::broadcast_keepalive)
-        .with(broadcasters::broadcast_movement)
-        .with(broadcasters::broadcast_velocity)
+        .with(player::check_crossed_chunks)
+        .with(player::broadcast_keepalive)
+        .with(entity::broadcast_movement)
+        .with(entity::broadcast_velocity)
         .with(entity::falling_block::spawn_falling_blocks)
-        .with(save::chunk_save)
+        .with(chunk_logic::chunk_save)
         .with(game::reset_bump_allocators)
         .with(game::increment_tick_count)
-        .with(time::increment_time)
+        .with(util::increment_time)
         .with(entity::previous_position_velocity_reset) // should be at end
 }
