+++ server/util/src/lib.rs
@@ -12,8 +12,6 @@ mod chunk_entities;
 pub use chunk_entities::*;
 mod time;
 pub use time::*;
-mod task;
-pub use task::*;
 mod load;
 pub use load::*;
