+++ server/weather/Cargo.toml
@@ -5,3 +5,9 @@ authors = ["caelunshun <caelunshun@gmail.com>"]
 edition = "2018"
 
 [dependencies]
+feather-core = { path = "../../core" }
+feather-server-types = { path = "../types" }
+feather-server-util = { path = "../util" }
+
+fecs = { path = "/home/caelum/dev/fecs" }
+rand = "0.7"
